🔨 Refactor: Modify some data sturcutre and make it able to work in docker-compose
