+++ client/package.json
@@ -35,5 +35,5 @@
     "reactstrap": "^6.5.0",
     "styled-components": "^4.1.2"
   },
-  "proxy": "http://localhost:8081"
+  "proxy": "http://backend:9999"
 }
